+++ src/web/components/table/__tests__/TableHead.test.tsx
@@ -5,7 +5,7 @@
 
 import {describe, test, expect, testing} from '@gsa/testing';
 import {screen, rendererWith, fireEvent, RendererOptions} from 'web/testing';
-import TableHead from 'web/components/table/Head';
+import TableHead from 'web/components/table/TableHead';
 
 const rendererWithTableHeader = (options: RendererOptions = {}) => {
   const {render, ...other} = rendererWith(options);
