+++ src/web/pages/cpes/Table.jsx
@@ -5,8 +5,8 @@
 
 import React from 'react';
 import {_, _l} from 'gmp/locale/lang';
-import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
+import TableHead from 'web/components/table/TableHead';
 import TableRow from 'web/components/table/TableRow';
 import {createEntitiesFooter} from 'web/entities/Footer';
 import {withEntitiesHeader} from 'web/entities/Header';
