+++ src/web/pages/hosts/Identifiers.jsx
@@ -12,9 +12,9 @@ import DetailsLink from 'web/components/link/DetailsLink';
 import TableBody from 'web/components/table/Body';
 import TableData from 'web/components/table/Data';
 import TableFooter from 'web/components/table/Footer';
-import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
 import Table from 'web/components/table/StripedTable';
+import TableHead from 'web/components/table/TableHead';
 import TableRow from 'web/components/table/TableRow';
 import DetailsBlock from 'web/entity/Block';
 import useTranslation from 'web/hooks/useTranslation';
