+++ src/web/pages/portlists/PortRangesTable.tsx
@@ -8,9 +8,9 @@ import {isDefined} from 'gmp/utils/identity';
 import DeleteIcon from 'web/components/icon/DeleteIcon';
 import TableBody from 'web/components/table/Body';
 import TableData from 'web/components/table/Data';
-import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
 import Table from 'web/components/table/StripedTable';
+import TableHead from 'web/components/table/TableHead';
 import TableRow from 'web/components/table/TableRow';
 import useTranslation from 'web/hooks/useTranslation';
