+++ src/web/pages/reportconfigs/Dialog.jsx
@@ -19,9 +19,9 @@ import TextField from 'web/components/form/TextField';
 import YesNoRadio from 'web/components/form/YesNoRadio';
 import TableBody from 'web/components/table/Body';
 import TableData from 'web/components/table/Data';
-import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
 import Table from 'web/components/table/Table';
+import TableHead from 'web/components/table/TableHead';
 import TableRow from 'web/components/table/TableRow';
 import PropTypes from 'web/utils/PropTypes';
 import withGmp from 'web/utils/withGmp';
