+++ src/web/pages/reports/AuditReportsTable.jsx
@@ -7,8 +7,8 @@ import React from 'react';
 import {_l} from 'gmp/locale/lang';
 import {isDefined} from 'gmp/utils/identity';
 import ComplianceState from 'web/components/label/ComplianceState';
-import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
+import TableHead from 'web/components/table/TableHead';
 import TableRow from 'web/components/table/TableRow';
 import {createEntitiesFooter} from 'web/entities/Footer';
 import {createEntitiesTable} from 'web/entities/Table';
