+++ src/web/pages/reports/details/CvesTable.jsx
@@ -11,8 +11,8 @@ import Divider from 'web/components/layout/Divider';
 import CveLink from 'web/components/link/CveLink';
 import DetailsLink from 'web/components/link/DetailsLink';
 import TableData from 'web/components/table/Data';
-import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
+import TableHead from 'web/components/table/TableHead';
 import TableRow from 'web/components/table/TableRow';
 import {createEntitiesTable} from 'web/entities/Table';
 import PropTypes from 'web/utils/PropTypes';
