+++ src/web/pages/reports/details/OperatingSystemsTable.jsx
@@ -12,8 +12,8 @@ import OsIcon from 'web/components/icon/OsIcon';
 import IconDivider from 'web/components/layout/IconDivider';
 import Link from 'web/components/link/Link';
 import TableData from 'web/components/table/Data';
-import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
+import TableHead from 'web/components/table/TableHead';
 import TableRow from 'web/components/table/TableRow';
 import {createEntitiesTable} from 'web/entities/Table';
 import PropTypes from 'web/utils/PropTypes';
