+++ src/web/pages/reports/details/PortsTable.jsx
@@ -7,8 +7,8 @@ import React from 'react';
 import {_, _l} from 'gmp/locale/lang';
 import SeverityBar from 'web/components/bar/SeverityBar';
 import TableData from 'web/components/table/Data';
-import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
+import TableHead from 'web/components/table/TableHead';
 import TableRow from 'web/components/table/TableRow';
 import {createEntitiesTable} from 'web/entities/Table';
 import PropTypes from 'web/utils/PropTypes';
