+++ src/web/pages/reports/details/TlsCertificatesTable.jsx
@@ -10,8 +10,8 @@ import DateTime from 'web/components/date/DateTime';
 import {DownloadIcon} from 'web/components/icon';
 import Link from 'web/components/link/Link';
 import TableData from 'web/components/table/Data';
-import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
+import TableHead from 'web/components/table/TableHead';
 import TableRow from 'web/components/table/TableRow';
 import RowDetailsToggle from 'web/entities/RowDetailsToggle';
 import {createEntitiesTable} from 'web/entities/Table';
