+++ src/web/pages/results/Table.jsx
@@ -8,8 +8,8 @@ import {_, _l} from 'gmp/locale/lang';
 import {SolutionTypeSvgIcon} from 'web/components/icon';
 import Layout from 'web/components/layout/Layout';
 import Sort from 'web/components/sortby/SortBy';
-import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
+import TableHead from 'web/components/table/TableHead';
 import TableRow from 'web/components/table/TableRow';
 import {createEntitiesFooter} from 'web/entities/Footer';
 import {withEntitiesHeader} from 'web/entities/Header';
