+++ src/web/pages/roles/DetailsPage.jsx
@@ -22,9 +22,9 @@ import Tabs from 'web/components/tab/Tabs';
 import TabsContainer from 'web/components/tab/TabsContainer';
 import TableBody from 'web/components/table/Body';
 import TableData from 'web/components/table/Data';
-import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
 import Table from 'web/components/table/StripedTable';
+import TableHead from 'web/components/table/TableHead';
 import TableRow from 'web/components/table/TableRow';
 import EntityNameTableData from 'web/entities/EntityNameTableData';
 import EntitiesTab from 'web/entity/EntitiesTab';
