+++ src/web/pages/scanconfigs/EditNvtDetailsDialog.jsx
@@ -18,10 +18,10 @@ import Link from 'web/components/link/Link';
 import Loading from 'web/components/loading/Loading';
 import TableBody from 'web/components/table/Body';
 import TableData from 'web/components/table/Data';
-import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
 import SimpleTable from 'web/components/table/SimpleTable';
 import Table from 'web/components/table/StripedTable';
+import TableHead from 'web/components/table/TableHead';
 import TableRow from 'web/components/table/TableRow';
 import useTranslation from 'web/hooks/useTranslation';
 import NvtPreference from 'web/pages/nvts/NvtPreference';
