+++ src/web/pages/scanconfigs/NvtFamilies.jsx
@@ -18,9 +18,9 @@ import Divider from 'web/components/layout/Divider';
 import Section from 'web/components/section/Section';
 import TableBody from 'web/components/table/Body';
 import TableData from 'web/components/table/Data';
-import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
 import Table from 'web/components/table/StripedTable';
+import TableHead from 'web/components/table/TableHead';
 import TableRow from 'web/components/table/TableRow';
 import useTranslation from 'web/hooks/useTranslation';
 import Trend from 'web/pages/scanconfigs/Trend';
