+++ src/web/pages/user-settings/DefaultsSettings.tsx
@@ -11,9 +11,9 @@ import Select from 'web/components/form/Select';
 import Layout from 'web/components/layout/Layout';
 import DetailsLink from 'web/components/link/DetailsLink';
 import TableBody from 'web/components/table/Body';
-import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
 import StripedTable from 'web/components/table/StripedTable';
+import TableHead from 'web/components/table/TableHead';
 import TableRow from 'web/components/table/TableRow';
 import useCapabilities from 'web/hooks/useCapabilities';
 import useShallowEqualSelector from 'web/hooks/useShallowEqualSelector';
